Add files via upload
The focus of this project is determining number of AGV that need to use for production. For determining workload location, mission and other variables of AGV is observed. By using machine algorithm in Python, we implement classification and regression to our codes for gathering information from data. As a result, we examine the reliability of codes in term of regression and classification performance metrics.